requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
jupyter
faiss-cpu
sentence-transformers
langchain
absl-py
transformers